BZ-1166011 - SessionInfo primary key is Integer

+++ kie-api/src/main/java/org/kie/api/task/model/TaskData.java
@@ -47,7 +47,7 @@ public interface TaskData extends Externalizable {
     
     String getDeploymentId();
 
-	int getProcessSessionId();
+	long getProcessSessionId();
 
     String getDocumentType();
